Initial experiment with Web Component in Elm
This was actually done over a few months of hack nights, playing with
Web Components, polyfills, and Elm.

I've left in the code to support Web Components v0, but will be deleting it
shortly, as there's not likely to be any use for it in the future.

This is working well in Chrome (57) and Safari (10.1). It's working in
Firefox (52), but not well -- CSS is leaking into the component, and the
main reason to use components is to prevent that.